+++ src/brew_tools/brew_maths.py
@@ -244,3 +244,38 @@ def fg_from_attenuation(og, attenuation):
     """
     fg = (1.0 - (attenuation / 100.0)) * to_plato(og)
     return to_sg(fg)
+
+
+def adjust_gravity_volume(vol, og, ng):
+    """
+    Returns the new volume needed to achieve the desired new gravity.
+    This is unit independent and the return value can be used for liters
+    and or gallons.
+
+    New Volume = (Volume * original Gravity) / new Gravity
+
+    :arg vol: Original volume of wort
+    :arg og: The current gravity of the wort
+    :arg ng: The desired gravity of the wort
+    :returns: The amount to adjust the wort volume by
+    """
+    og = (og - 1) * 1000
+    ng = (ng - 1) * 1000
+    return (vol * og) / ng
+
+
+def adjust_volume_gravity(vol, og, new_vol):
+    """
+    Calculate the new gravity after boil off or dilution to ``new_vol``
+    This is unit independent and the volume can be used for liters
+    and or gallons.
+
+    Ending Gravity = (Beginning Volume * Beginning Gravity) / End Volume
+
+    :arg vol: Original volume of wort
+    :arg og: The current gravity of the wort
+    :arg new_vol: The new volume of the wort
+    :returns: The new gravity after boiloff or dilution
+    """
+    og = (og - 1) * 1000
+    return 1 + ((vol * og) / new_vol) / 1000